+++ lib/xe_currency.rb
@@ -2,4 +2,5 @@
 
 require 'money'
 require 'money/bank/xe_currency'
+require 'xe_currency/client'
 require 'xe_currency/version'
