+++ spec/money/bank/xe_currency_spec.rb
@@ -3,7 +3,11 @@
 require 'spec_helper'
 
 describe Money::Bank::XeCurrency do
+  let(:client) { instance_double(::XeCurrency::Client) }
+
   before do
+    allow(::XeCurrency::Client).to receive(:new).and_return(client)
+
     Money.default_bank = Money::Bank::XeCurrency.new(
       Money::RatesStore::Memory.new,
       account_api_id: 'fake',
@@ -12,10 +16,8 @@ describe Money::Bank::XeCurrency do
   end
 
   it 'converts rates' do
-    stub_request(:get, 'https://xecdapi.xe.com/v1/convert_from.json?from=JPY&to=USD')
-      .to_return(body: get_response('jpy_to_usd.json'))
-    stub_request(:get, 'https://xecdapi.xe.com/v1/convert_from.json?from=USD&to=JPY')
-      .to_return(body: get_response('usd_to_jpy.json'))
+    expect(client).to receive(:fetch_rate).with('JPY', 'USD').and_return(BigDecimal('0.0077759754'))
+    expect(client).to receive(:fetch_rate).with('USD', 'JPY').and_return(BigDecimal('128.6012304362'))
 
     money = Money.new(1_00, 'JPY')
     expect(money.exchange_to(:USD).to_s).to eq('0.78')
@@ -25,8 +27,7 @@ describe Money::Bank::XeCurrency do
   end
 
   it 'raises an error with a given message' do
-    stub_request(:get, 'https://xecdapi.xe.com/v1/convert_from.json?from=JPY&to=USD')
-      .to_return(status: 401, body: get_response('error_bad_credentials.json'))
+    expect(client).to receive(:fetch_rate).and_raise(::XeCurrency::Client::FetchError, 'Bad credentials')
 
     money = Money.new(1_00, 'JPY')
     expect { money.exchange_to(:USD) }.to raise_error(Money::Bank::XeCurrency::XeCurrencyFetchError, 'Bad credentials')
